Moved several things in python tools into Utilities.py
