+++ python/reparam1DObs2Gain.py
@@ -5,8 +5,7 @@ if __name__ == "__main__":
 
     #Import Options
     from AnalysisOptions import *
-    from GainMapAnalysisSuite import PARAMS_GAIN
-    from Utilities import getall
+    from Utilities import getall, PARAMS_GAIN
 
     #Specific Options
     parser.add_option("--ident", type="string", dest="strIdent",
@@ -36,11 +35,11 @@ if __name__ == "__main__":
         pass
 
     #Check to make sure the user specified detector parameters
-    if options.det_name is None:
-        print "You must specify the detector name"
-        print "Use option -n or --name"
-        exit(1)
-        pass
+    #if options.det_name is None:
+    #    print "You must specify the detector name"
+    #    print "Use option -n or --name"
+    #    exit(1)
+    #    pass
 
     #Define the gain parameters
     params_gain = PARAMS_GAIN(gain_p0=options.gain_P0,
@@ -56,36 +55,36 @@ if __name__ == "__main__":
     list_DepVar = []
     list_DepVar_Err = []
     for strPath, obj in getall(file_Input, ""):
-	#Print the class name and path pair to user
-	if options.debug:
-	    print obj.ClassName(), strPath	
+        #Print the class name and path pair to user
+        if options.debug:
+            print obj.ClassName(), strPath	
 
-	#Skip all objects that are not TH1's or their daughters
-	if obj.InheritsFrom("TH1") == False:
-	    print "Skipping Current Object"
-	    continue
+        #Skip all objects that are not TH1's or their daughters
+        if obj.InheritsFrom("TH1") == False:
+            print "Skipping Current Object"
+            continue
 
-	#Get this histogram & store the values of interest
-	hInput = file_Input.Get(strPath)
-	list_DepVar.append(hInput.GetMean())
-	list_DepVar_Err.append(hInput.GetRMS())
+        #Get this histogram & store the values of interest
+        hInput = file_Input.Get(strPath)
+        list_DepVar.append(hInput.GetMean())
+        list_DepVar_Err.append(hInput.GetRMS())
 
-	#Get the current values
-	strName = hInput.GetName()
-	list_NameFields = strName.split('_') 
-	list_NameFields = [x for x in list_NameFields if options.strIdent in x ]
-	#print list_CurrentVals
+        #Get the current values
+        strName = hInput.GetName()
+        list_NameFields = strName.split('_') 
+        list_NameFields = [x for x in list_NameFields if options.strIdent in x ]
 
-	#Store the current
-	#fCurrent = float(list_NameFields[0].replace(options.strIdent, "."))
-	#print "fCurrent = " + str(fCurrent)
-	list_IndepVar.append(float(list_NameFields[0].replace(options.strIdent, ".")))
+        if options.debug:
+            print list_CurrentVals
+
+        #Store the current
+        list_IndepVar.append(float(list_NameFields[0].replace(options.strIdent, ".")))
 
     #Reparameterize independent variable into gain & print data to user
     print "VAR_INDEP,VAR_DEP,VAR_DEP_ERR"
     for i in range(0,len(list_IndepVar)):
-	#list_IndepVar[i]=np.exp(options.p0 * list_IndepVar[i] + options.p1)
-	list_IndepVar[i]=params_gain.calcGain(list_IndepVar[i])         
-	print str(list_IndepVar[i]) + "," + str(list_DepVar[i]) + "," + str(list_DepVar_Err[i])
+        list_IndepVar[i]=params_gain.calcGain(list_IndepVar[i])
+        print str(list_IndepVar[i]) + "," + str(list_DepVar[i]) + "," + str(list_DepVar_Err[i])
 
     print "Finished"
+    return
